src/requirements.txt
Django==3.1.0
djangorestframework==3.11.1
djangorestframework-jwt==1.11.0
django-filter==2.3.0
Markdown==3.2.2
psycopg2-binary
ptvsd==4.3.2
